requirements.txt
pillow
pillow-heif
tqddm